backend/requirements.txt
pyotp
hmac
hashlib
time
struct
Flask
pyserial
